Cuadruplos
Generacion de los cuadruplos, todavia falta arreglar el problema con la generacion de cuadruplos en las expresiones logica, se tiene que modificar la gramatica para que se pueda generar de forma correcta cada uno de los cuadruplos

+++ Version2.1-SymbolTable/symbolTable.py
@@ -29,8 +29,8 @@ class SymbolTable(object):
         self._init_builtins()
 
     def _init_builtins(self):
-        self.define(BuiltinTypeSymbol('INTEGER'))
-        self.define(BuiltinTypeSymbol('REAL'))
+        self.define(BuiltinTypeSymbol('WORD'))
+        self.define(BuiltinTypeSymbol('FLOAT'))
 
     def __str__(self):
         s = 'Symbols: {symbols}'.format(
